# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  ent.CheckIn:
    properties:
      checkin_date:
        description: CheckinDate holds the value of the "checkin_date" field.
        type: string
      edges:
        $ref: '#/definitions/ent.CheckInEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the CheckInQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.CheckInEdges:
    properties:
      checkouts:
        $ref: '#/definitions/ent.Checkout'
        description: Checkouts holds the value of the checkouts edge.
        type: object
      counter:
        $ref: '#/definitions/ent.CounterStaff'
        description: Counter holds the value of the counter edge.
        type: object
      customer:
        $ref: '#/definitions/ent.Customer'
        description: Customer holds the value of the customer edge.
        type: object
      dataroom:
        $ref: '#/definitions/ent.DataRoom'
        description: Dataroom holds the value of the dataroom edge.
        type: object
      reserveroom:
        $ref: '#/definitions/ent.ReserveRoom'
        description: Reserveroom holds the value of the reserveroom edge.
        type: object
    type: object
  ent.Checkout:
    properties:
      checkout_date:
        description: CheckoutDate holds the value of the "checkout_date" field.
        type: string
      edges:
        $ref: '#/definitions/ent.CheckoutEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the CheckoutQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.CheckoutEdges:
    properties:
      checkins:
        $ref: '#/definitions/ent.CheckIn'
        description: Checkins holds the value of the checkins edge.
        type: object
      counterstaffs:
        $ref: '#/definitions/ent.CounterStaff'
        description: Counterstaffs holds the value of the counterstaffs edge.
        type: object
      statuss:
        $ref: '#/definitions/ent.Status'
        description: Statuss holds the value of the statuss edge.
        type: object
    type: object
  ent.CounterStaff:
    properties:
      edges:
        $ref: '#/definitions/ent.CounterStaffEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the CounterStaffQuery when eager-loading is set.
        type: object
      email:
        description: Email holds the value of the "email" field.
        type: string
      id:
        description: ID of the ent.
        type: integer
      name:
        description: Name holds the value of the "name" field.
        type: string
      password:
        description: Password holds the value of the "password" field.
        type: string
    type: object
  ent.CounterStaffEdges:
    properties:
      checkins:
        description: Checkins holds the value of the checkins edge.
        items:
          $ref: '#/definitions/ent.CheckIn'
        type: array
      checkouts:
        description: Checkouts holds the value of the checkouts edge.
        items:
          $ref: '#/definitions/ent.Checkout'
        type: array
    type: object
  ent.Customer:
    properties:
      edges:
        $ref: '#/definitions/ent.CustomerEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the CustomerQuery when eager-loading is set.
        type: object
      email:
        description: Email holds the value of the "email" field.
        type: string
      id:
        description: ID of the ent.
        type: integer
      name:
        description: Name holds the value of the "name" field.
        type: string
      password:
        description: Password holds the value of the "password" field.
        type: string
    type: object
  ent.CustomerEdges:
    properties:
      checkins:
        description: Checkins holds the value of the checkins edge.
        items:
          $ref: '#/definitions/ent.CheckIn'
        type: array
      fixs:
        description: Fixs holds the value of the fixs edge.
        items:
          $ref: '#/definitions/ent.FixRoom'
        type: array
      reserves:
        description: Reserves holds the value of the reserves edge.
        items:
          $ref: '#/definitions/ent.ReserveRoom'
        type: array
    type: object
  ent.DataRoom:
    properties:
      edges:
        $ref: '#/definitions/ent.DataRoomEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the DataRoomQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
      price:
        description: Price holds the value of the "price" field.
        type: number
      roomnumber:
        description: Roomnumber holds the value of the "roomnumber" field.
        type: string
    type: object
  ent.DataRoomEdges:
    properties:
      checkins:
        description: Checkins holds the value of the checkins edge.
        items:
          $ref: '#/definitions/ent.CheckIn'
        type: array
      details:
        description: Details holds the value of the details edge.
        items:
          $ref: '#/definitions/ent.FurnitureDetail'
        type: array
      fixs:
        description: Fixs holds the value of the fixs edge.
        items:
          $ref: '#/definitions/ent.FixRoom'
        type: array
      promotion:
        $ref: '#/definitions/ent.Promotion'
        description: Promotion holds the value of the promotion edge.
        type: object
      reserves:
        description: Reserves holds the value of the reserves edge.
        items:
          $ref: '#/definitions/ent.ReserveRoom'
        type: array
      statusroom:
        $ref: '#/definitions/ent.StatusRoom'
        description: Statusroom holds the value of the statusroom edge.
        type: object
      typeroom:
        $ref: '#/definitions/ent.TypeRoom'
        description: Typeroom holds the value of the typeroom edge.
        type: object
    type: object
  ent.FixRoom:
    properties:
      edges:
        $ref: '#/definitions/ent.FixRoomEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the FixRoomQuery when eager-loading is set.
        type: object
      fix_detail:
        description: FixDetail holds the value of the "fix_detail" field.
        type: string
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.FixRoomEdges:
    properties:
      customer:
        $ref: '#/definitions/ent.Customer'
        description: Customer holds the value of the customer edge.
        type: object
      furnitureDetail:
        $ref: '#/definitions/ent.FurnitureDetail'
        description: FurnitureDetail holds the value of the furnitureDetail edge.
        type: object
      room:
        $ref: '#/definitions/ent.DataRoom'
        description: Room holds the value of the room edge.
        type: object
    type: object
  ent.Furniture:
    properties:
      edges:
        $ref: '#/definitions/ent.FurnitureEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the FurnitureQuery when eager-loading is set.
        type: object
      furniture_name:
        description: FurnitureName holds the value of the "furniture_name" field.
        type: string
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.FurnitureDetail:
    properties:
      date_add:
        description: DateAdd holds the value of the "date_add" field.
        type: string
      edges:
        $ref: '#/definitions/ent.FurnitureDetailEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the FurnitureDetailQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.FurnitureDetailEdges:
    properties:
      fixs:
        description: Fixs holds the value of the fixs edge.
        items:
          $ref: '#/definitions/ent.FixRoom'
        type: array
      furnitures:
        $ref: '#/definitions/ent.Furniture'
        description: Furnitures holds the value of the furnitures edge.
        type: object
      rooms:
        $ref: '#/definitions/ent.DataRoom'
        description: Rooms holds the value of the rooms edge.
        type: object
      types:
        $ref: '#/definitions/ent.FurnitureType'
        description: Types holds the value of the types edge.
        type: object
    type: object
  ent.FurnitureEdges:
    properties:
      details:
        description: Details holds the value of the details edge.
        items:
          $ref: '#/definitions/ent.FurnitureDetail'
        type: array
    type: object
  ent.FurnitureType:
    properties:
      edges:
        $ref: '#/definitions/ent.FurnitureTypeEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the FurnitureTypeQuery when eager-loading is set.
        type: object
      furniture_type:
        description: FurnitureType holds the value of the "furniture_type" field.
        type: string
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.FurnitureTypeEdges:
    properties:
      details:
        description: Details holds the value of the details edge.
        items:
          $ref: '#/definitions/ent.FurnitureDetail'
        type: array
    type: object
  ent.Promotion:
    properties:
      discount:
        description: Discount holds the value of the "discount" field.
        type: number
      edges:
        $ref: '#/definitions/ent.PromotionEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the PromotionQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
      promotion_name:
        description: PromotionName holds the value of the "promotion_name" field.
        type: string
    type: object
  ent.PromotionEdges:
    properties:
      datarooms:
        description: Datarooms holds the value of the datarooms edge.
        items:
          $ref: '#/definitions/ent.DataRoom'
        type: array
      reserves:
        description: Reserves holds the value of the reserves edge.
        items:
          $ref: '#/definitions/ent.ReserveRoom'
        type: array
    type: object
  ent.ReserveRoom:
    properties:
      date_out:
        description: DateOut holds the value of the "date_out" field.
        type: string
      edges:
        $ref: '#/definitions/ent.ReserveRoomEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the ReserveRoomQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
      net_price:
        description: NetPrice holds the value of the "net_price" field.
        type: number
      reserve_date:
        description: ReserveDate holds the value of the "reserve_date" field.
        type: string
    type: object
  ent.ReserveRoomEdges:
    properties:
      checkins:
        description: Checkins holds the value of the checkins edge.
        items:
          $ref: '#/definitions/ent.CheckIn'
        type: array
      customer:
        $ref: '#/definitions/ent.Customer'
        description: Customer holds the value of the customer edge.
        type: object
      promotion:
        $ref: '#/definitions/ent.Promotion'
        description: Promotion holds the value of the promotion edge.
        type: object
      room:
        $ref: '#/definitions/ent.DataRoom'
        description: Room holds the value of the room edge.
        type: object
      status:
        $ref: '#/definitions/ent.StatusReserve'
        description: Status holds the value of the status edge.
        type: object
    type: object
  ent.Status:
    properties:
      description:
        description: Description holds the value of the "description" field.
        type: string
      edges:
        $ref: '#/definitions/ent.StatusEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the StatusQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.StatusEdges:
    properties:
      checkouts:
        description: Checkouts holds the value of the checkouts edge.
        items:
          $ref: '#/definitions/ent.Checkout'
        type: array
    type: object
  ent.StatusReserve:
    properties:
      edges:
        $ref: '#/definitions/ent.StatusReserveEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the StatusReserveQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
      status_name:
        description: StatusName holds the value of the "status_name" field.
        type: string
    type: object
  ent.StatusReserveEdges:
    properties:
      reserves:
        description: Reserves holds the value of the reserves edge.
        items:
          $ref: '#/definitions/ent.ReserveRoom'
        type: array
    type: object
  ent.StatusRoom:
    properties:
      edges:
        $ref: '#/definitions/ent.StatusRoomEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the StatusRoomQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
      status_name:
        description: StatusName holds the value of the "status_name" field.
        type: string
    type: object
  ent.StatusRoomEdges:
    properties:
      datarooms:
        description: Datarooms holds the value of the datarooms edge.
        items:
          $ref: '#/definitions/ent.DataRoom'
        type: array
    type: object
  ent.TypeRoom:
    properties:
      edges:
        $ref: '#/definitions/ent.TypeRoomEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the TypeRoomQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
      type_name:
        description: TypeName holds the value of the "type_name" field.
        type: string
    type: object
  ent.TypeRoomEdges:
    properties:
      datarooms:
        description: Datarooms holds the value of the datarooms edge.
        items:
          $ref: '#/definitions/ent.DataRoom'
        type: array
    type: object
  gin.H:
    additionalProperties: true
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for SUT SE 2563
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: SUT SE Example API
  version: "1.0"
paths:
...
  /Customers:
    get:
      description: list Customer entities
      operationId: list-Customer
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ent.Customer'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: List Customer entities
  /Customers/{id}:
    get:
      description: get Customer by ID
      operationId: get-Customer
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.Customer'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get a Customer entity by ID
...
  /checkouts:
    get:
      description: list checkout entities
      operationId: list-checkout
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ent.Checkout'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: List checkout entities
    post:
      consumes:
      - application/json
      description: Create checkout
      operationId: create-checkout
      parameters:
      - description: Checkout entity
        in: body
        name: checkout
        required: true
        schema:
          $ref: '#/definitions/ent.Checkout'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.Checkout'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Create checkout
  /checkouts/{id}:
    delete:
      description: get checkout by ID
      operationId: delete-checkout
      parameters:
      - description: Checkout ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Delete a checkout entity by ID
    get:
      description: get checkout by ID
      operationId: get-checkout
      parameters:
      - description: Checkout ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.Checkout'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get a checkout entity by ID
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  BasicAuth:
    type: basic
  OAuth2AccessCode:
    authorizationUrl: https://example.com/oauth/authorize
    flow: accessCode
    scopes:
      admin: ' Grants read and write access to administrative information'
    tokenUrl: https://example.com/oauth/token
    type: oauth2
  OAuth2Application:
    authorizationUrl: ""
    flow: application
    scopes:
      admin: ' Grants read and write access to administrative information'
      write: ' Grants write access'
    tokenUrl: https://example.com/oauth/token
    type: oauth2
  OAuth2Implicit:
    authorizationUrl: https://example.com/oauth/authorize
    flow: implicit
    scopes:
      admin: ' Grants read and write access to administrative information'
      write: ' Grants write access'
    type: oauth2
  OAuth2Password:
    authorizationUrl: ""
    flow: password
    scopes:
      admin: ' Grants read and write access to administrative information'
      read: ' Grants read access'
      write: ' Grants write access'
    tokenUrl: https://example.com/oauth/token
    type: oauth2
swagger: "2.0"